fix: move components to app/lib
