+++ packages/apps/frequency-wallet-proxy/package.json
@@ -16,6 +16,7 @@
   },
   "devDependencies": {
     "@frequency-control-panel/eslint-config": "workspace:*",
+    "@iconify/types": "^2.0.0",
     "@playwright/test": "^1.28.1",
     "@sveltejs/adapter-static": "^3.0.1",
     "@sveltejs/kit": "^2.0.0",
@@ -37,7 +38,10 @@
   },
   "type": "module",
   "dependencies": {
-    "@frequency-control-panel/components": "workspace:^",
-    "@polkadot/extension-inject": "^0.46.6"
+    "@frequency-control-panel/utils": "workspace:^",
+    "@iconify/icons-ic": "^1.2.13",
+    "@iconify/svelte": "^3.1.6",
+    "@polkadot/extension-inject": "^0.46.6",
+    "unplugin-icons": "^0.18.2"
   }
 }
