+++ packages/apps/frequency-wallet-proxy/src/lib/components/WalletButton.svelte
