+++ packages/apps/frequency-wallet-proxy/src/lib/components/WalletSelector.svelte
