+++ packages/apps/frequency-wallet-proxy/src/lib/components/index.ts
