+++ packages/apps/frequency-wallet-proxy/src/lib/icons/FrequencyGradient.svelte
