+++ packages/apps/frequency-wallet-proxy/src/lib/icons/Polkadot.svelte
