+++ packages/apps/frequency-wallet-proxy/src/lib/icons/SubWallet.svelte
