+++ packages/apps/frequency-wallet-proxy/src/lib/icons/TalismanRedIcon.svelte
