--- packages/components/.eslintignore
@@ -1,14 +1,0 @@
-.DS_Store
-node_modules
-/build
-/.svelte-kit
-/package
-/dist
-/src/lib/icons
-.env
-.env.*
-!.env.example
-
-# Ignore files for PNPM, NPM and YARN
-pnpm-lock.yaml
-package-lock.json
