--- packages/components/.eslintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['@frequency-control-panel/eslint-config/index.js'],
-};
