--- packages/components/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": ["./.svelte-kit/tsconfig.json", "../../tsconfig.base.json"],
-  "compilerOptions": {
-    "allowJs": true,
-    "checkJs": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "strict": true,
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext"
-  }
-}
