--- packages/components/vite.config.ts
@@ -1,15 +1,0 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vitest/config';
-import Icons from 'unplugin-icons/vite';
-
-export default defineConfig({
-  plugins: [
-    sveltekit(),
-    Icons({
-      compiler: 'svelte',
-    }),
-  ],
-  test: {
-    include: ['src/**/*.{test,spec}.{js,ts}'],
-  },
-});
